use trait for keyword search
